Revert "refactor: update user repository to interface and adjust service constructors"
This reverts commit 37f375a2c070e21b7c0b517df352d54f829ee249.
