--- internal/repository/di/providers.go
@@ -1,10 +1,0 @@
-package di
-
-import (
-	"github.com/scienceandcode/habits-todo-backend/internal/repository"
-	"gorm.io/gorm"
-)
-
-func ProvideUserRepository(db *gorm.DB) repository.UserRepository {
-	return repository.NewUserRepository(db)
-}
